Fix all classes to now actually run api

+++ Backend/src/main/java/com/backend/Java_Backend/Services/ThreadParticipantService.java
@@ -18,7 +18,7 @@ public class ThreadParticipantService {
         this.participantRepo = participantRepo;
     }
 
-    public ThreadParticipant addParticipant(UUID threadID, Long userId) {
+    public ThreadParticipant addParticipant(UUID threadID, int userId) {
         ThreadParticipantId id = new ThreadParticipantId(threadID, userId);
         ThreadParticipant participant = new ThreadParticipant(id);
         return participantRepo.save(participant);
